--- Problem Solving Practice/src/LeetCode/Medium/ProductOfArrayExceptSelf/ProductOfArrayExceptSelf.java
@@ -1,33 +1,0 @@
-package LeetCode.Medium.ProductOfArrayExceptSelf;
-
-import java.util.Arrays;
-
-public class ProductOfArrayExceptSelf {
-    public static void main(String[] args) {
-        int[] nums1 = {1, 2, 3, 4};
-        int[] result = new Solution().productExceptSelf(nums1);
-        System.out.println(Arrays.toString(result));
-    }
-}
-
-class Solution {
-    public int[] productExceptSelf(int[] nums) {
-        int n = nums.length;
-        int[] leftProduct = new int[n];
-        int[] rightProduct = new int[n];
-        int[] result = new int[n];
-        leftProduct[0] = 1;
-        for (int i = 1; i < n; i++) {
-            leftProduct[i] = leftProduct[i - 1] * nums[i - 1];
-        }
-        rightProduct[n - 1] = 1;
-        for (int i = n - 2; i >= 0; i--) {
-            rightProduct[i] = rightProduct[i + 1] * nums[i + 1];
-        }
-        for (int i = 0; i < n; i++) {
-            result[i] = leftProduct[i] * rightProduct[i];
-        }
-
-        return result;
-    }
-}
