+++ Problem Solving Practice/src/LeetCode/TopInterview150/LongestCommonPrefix/14. Longest Common Prefix
