+++ Problem Solving Practice/src/LeetCode/TopInterview150/RemoveElement/RemoveElement.java
@@ -1,0 +1,26 @@
+package LeetCode.TopInterview150.RemoveElement;
+
+public class RemoveElement {
+    public static void main(String[] args) {
+        int[] nums = {3,2,2,3};
+        int val = 3;
+        Solution solution = new Solution();
+        System.out.println(solution.removeElement(nums, val));
+
+    }
+}
+class Solution {
+    public int removeElement(int[] nums, int val) {
+        int k = 0;
+        for (int i = 0; i < nums.length; i++) {
+            if (nums[i] != val) {
+                nums[k] = nums[i];
+                k++;
+            }
+        }
+        for (int i = k; i < nums.length; i++) {
+            nums[i] = 0;
+        }
+        return k;
+    }
+}
