+++ Problem Solving Practice/src/LeetCode/TopInterview150/RotateArray/RotateArray.java
@@ -1,0 +1,34 @@
+package LeetCode.TopInterview150.RotateArray;
+
+import java.util.Arrays;
+
+public class RotateArray {
+    public static void main(String[] args) {
+        int[] nums = {1, 2, 3, 4, 5, 6, 7};
+        int[] resl = {6, 7, 1, 2, 3, 4, 5};
+        new Solution().rotate(nums, 2);
+        System.out.println(Arrays.toString(nums));
+    }
+
+}
+
+class Solution {
+
+    public void rotate(int[] nums, int k) {
+        int n = nums.length;
+        k = k % n;
+        reverse(nums, 0, n - 1);
+        reverse(nums, 0, k - 1);
+        reverse(nums, k, n - 1);
+    }
+
+    private void reverse(int[] nums, int start, int end) {
+        while (start < end) {
+            int temp = nums[start];
+            nums[start] = nums[end];
+            nums[end] = temp;
+            start++;
+            end--;
+        }
+    }
+}
